Debugged world.analytic_cycles increment, copy and past error :P
Move game clock tests to their own module.
